refactor: extract output formatters from runner.py to formatters.py
